update to fastdate=v0.2

+++ rbdc-pg/src/types/date.rs
@@ -21,7 +21,7 @@ impl Decode for fastdate::Date {
                     year: 2000,
                     day: 1,
                     mon: 1,
-                    offset: fastdate::offset_sec() as i64,
+                    offset: fastdate::offset_sec(),
                 };
                 let dt = {
                     if days < 0 {
@@ -52,7 +52,7 @@ impl Encode for fastdate::Date {
             day: self.day,
             mon: self.mon,
             year: self.year,
-            offset: fastdate::offset_sec() as i64,
+            offset: fastdate::offset_sec(),
         }
         .unix_timestamp_millis()
             - fastdate::DateTime {
@@ -63,7 +63,7 @@ impl Encode for fastdate::Date {
                 year: 2000,
                 day: 1,
                 mon: 1,
-            offset: fastdate::offset_sec() as i64,
+            offset: fastdate::offset_sec(),
             }
             .unix_timestamp_millis())
             / (86400 * 1000) as i64;
